logic in commands update

+++ src/main/java/ru/neginskiy/tm/command/ProjectDeleteCommand.java
@@ -9,37 +9,45 @@ public class ProjectDeleteCommand extends AbstractCommand {
 
     @Override
     public void execute() {
-        System.out.println("Please select project number to remove :");
         List<Project> projectList = getBootstrap().getProjectService().getAll();
+        if (projectList.size() == 0) {
+            System.out.println("Projects not found");
+            return;
+        }
+
+        System.out.println("Please select project number to remove :");
         int indexOfProject = 0;
         for (Project project : projectList) {
             System.out.printf("%-3s%s%s%n", indexOfProject++, " - ", project.getName());
         }
-        int projectNumber;
-        String id;
+
+        Project project;
         try {
-            projectNumber = Integer.parseInt(getBootstrap().readLine());
-            Project project = projectList.get(projectNumber);
+            int projectNumber = Integer.parseInt(getBootstrap().readLine());
+            project = projectList.get(projectNumber);
             if (project == null) {
                 System.out.println("Incorrect input, project not found");
                 return;
             }
-            id = project.getId();
         } catch (Exception e) {
             System.out.println("Incorrect input, project not found");
             return;
         }
-
+        String id = project.getId();
         getBootstrap().getProjectService().delete(id);
+        System.out.println("Project deleted");
+
         List<Task> taskList = getBootstrap().getTaskService().getAll();
+        if (taskList.size() == 0) {
+            System.out.println("Tasks not found");
+            return;
+        }
         for (Task task : taskList) {
             if (id.equals(task.getProjectId())) {
                 getBootstrap().getProjectService().delete(task.getId());
             }
         }
-        //getBootstrap().getTaskService().deleteByProjectId(id);
-
-        System.out.println("Project and nested tasks deleted");
+        System.out.println("Project's tasks deleted");
     }
 
     @Override
